plat/kvm: Introduce SMP support for kvm x86
This commit implements the following:
	* parsing of the MADT table, that must be found, using
`acpi_init`, before calling `smp_init` and the subsequent API
functions
	* initialization of the smp module; for now, only the
x2APIC mode is initialized, as it offers the ability to increase,
in the future, the number of enabled cores, it uses model-specific
registers, not memory-mapped registers and is faster than the xAPIC
mode; this is done by the `smp_init` function
	* getting the number of available cores, via `ukplat_lcpu_count`
	* getting the ID of the current running core, via
`ukplat_lcpu_id`
	* getting whether a core is the bootstrapping one, via
`ukplat_lcpu_is_bsp`
	* enabling multiple cores, via `ukplat_lcpu_start`; only the stack
pointer is mandatory. If the lcpuid array is NULL, num will also default
to the maximum number of logical CPUs in the system. If the entry point
is NULL, the LCPUs will be put in a halted state.
	* waiting for some (or all) LCPUs to reach a halted state, via
`ukplat_lcpu_wait`

Signed-off-by: cristian-vijelie <cristianvijelie@gmail.com>